Refactoring some odd legacy code

+++ System.cs
@@ -9,13 +9,13 @@ namespace Abyss_Call
         public abstract bool Requirements(Entity e);
         public List<Entity> EntityBucket { get; set; } = new List<Entity>();
 
-        public void Update(int deltaTime)
+        public void Update(double deltaTime)
         {
             foreach (Entity e in EntityBucket)
                 if (e.IsUpdatable)
                     UpdateEntity(e, deltaTime);
         }
-        protected virtual void UpdateEntity(Entity entity, int deltaTime)
+        protected virtual void UpdateEntity(Entity entity, double deltaTime)
         {
             //Debug.WriteLine(entity.Tag, "Updating: ");
         }
